Initial commit of the rough landing page.
Right now, we're almost done with the reworked landing page, which also includes the behind-the-scenes endpoint to get the list of available clusters to monitor.  It's hard-coded to use the Consul-based instance discovery implementation, which we've also implemented here.  Not sure if there's a way to let Turbine get configured by Archaius and then ask it for the instance list, but we can investigate that later.