Rip out Turbine.
Since there's some deficiencies -- namely you can't have it monitor a changing list of clusters after it's been initialized -- we're going to punt on using it for the moment.  We're probably going to end up writing something that can be configured like Turbine (so that configs can be copypasted) but then do some amalgamation of Turbine 2 observables stuff with a new streaming servlet.  Turbine 2 doesn't seem to lend itself to being embedded in the servlet fashion since it wants to run its own web server with RxNetty.