Rough cut of integrating Turbine 2.0.
We're now using some of the newer Turbine 2.0 code to more cleanly proxy/aggregate streams.  We still have to flesh out the discovery portion for Consul so that it can be dropped in cleanly.  Right now, single host proxying works well, but we're unable to really use/test the cluster aggregation mode until the discovery portion is well and truly done.  With doing all of this, we're also moving away from servlets and using native Dropwizard/Jersey resources.