We have working aggregated cluster streams and cluster discovery.
I separated out how to discover clusters (used to drive the landing page) and finding the streams that compromise those clusters (used on the monitoring page).  Stream discovery follows Turbine v2's approach of being an observable that emits up/down events, primarily because I lifted a lot of code/ideas for doing the stream aggregation from Turbine v2 itself.  We're now fully using Dropwizard/Jersey resources for API actions, and things should be pretty well handled to avoid weirdness.  The default implementations for cluster and stream discovery use Consul.  It uses all the default, (so it tries localhost:8500 for the agent) and it looks for services tagged with 'hystrix'.  I'll probably make that configurable in the future.