Reorganize and implement the request graph.
The request graph is intended to represent the relative request volume and show a rolling window of request rate, instead of the instantaneous measurement shown for the host/cluster.  We don't have the line graph portion settled yet, but the request volume 'bubble' is there.  We still need to add in the coloring the bubble and the error percentage value based on the actual error percentage.  We're getting there.