Rough cut of a unified discovery story.
We now have an interface to provide cluster discovery from a single class. We also have a way to define multiple discovery providers to be used at once.  We only have a static cluster config provider at the moment under this new approach, but porting the Consul support will come soon.  There's still work to do to actually wire the cluster streaming endpoint up on the UI side but it's a small amount of work.