Rejigged the JS.
I need to figure out a clean way to use the Gulp pipeline to generate a bundled JS file for both the landing page and the monitor page.  Right now, I can't just drop in the bundled JS because it's going to try and boot into the monitor application.  For now, we'll just duplicate shit so it works... but this needs to get fixed.