Basically converted over to SPA.
We've switched over to Webpack, and moved to a SPA approach.  We also switched from Chosen to react-select, which is a little nicer and avoids other external dependencies.  All of our JS is now NPM-managed.  We also upgraded to React 0.14, which meant fixing some deprecations.

The main remaining bit is adding a button/approach to go back to the stream selection screen.