Switch to react-router for routing.
I was doing a weird thing of functions on the root component being passed to children, which they would then call to have the root component switch states/views.  Using the router makes that generic, extensible, and actually allows people to copy/paste links and have them work in a different browser.