Don't compress.  Reduce!
Rather than trying to do compression for streaming data -- which seems to be nigh impossible with the built-in Java stuff -- we can get a lot of bang for our buck by reducing the number of property values we send to the browser.  A lot of them don't show up at all in the UI, and are purely informational and otherwise duplicate information you could get by simply looking at the configuration you have for the commands/thread pools themselves.

In the future, we could potentially use something like Protocol Buffers/Thrift which have a rigid schema and don't require key name, letting us get down to literally just the data being expressed.  I don't think we'll get much better than that, including with compression, if we want to stream real-time.