First rough past at sorting.
This is following the idea of maintaining sort order externally and then reordering all values according to that order every time we go to dispatch.  There's no concept of sort ordering now, so sorting by name looms naturally -- A to Z -- but sorting by RPS shows the heaviest things last, which is not the most intuitive order.  I'm not sure if we should have sort direction as a toggleable thing or just use a specific direction per sort key...