Rough support for finding ELBs.
This uses the AWS Java SDK to look for ELBs that are tagged, indicating they run a Hystrix-capable service, and it grabs all the backend instances and uses those.  It assumes the instance port for the first listener is the port to connect to.  This might be a problem for some, but this was the simplest solution I could come up with that allows ELBs to sort of self-define.

It can be a little slower than static clusters or Consul, because we have to make multiple API calls... but it's not terrible.