Add a timeout to individual streams.
For whatever reason, once RxNetty establishes a connection to an SSE endpoint, it won't throw an error if that endpoint goes away.  By adding a timeout to the content observable, we can generate an error that propagates back to our retry handler, which should then trigger a retry which will try to reconnect.  The timeout is set at 10 seconds, which should suffice given that the normal interval for updates is 1 second.