Track when a value was received.
In order to figure out when commands / thread pools should be marked stale, we now track the timestamp of when we received a value for an aggregator.  This gets passed all the way to the MonitorPanel component.  When complete, the monitor panel will use that value to visually indicate that the command / thread pool being displayed is stale.  This may be because a cluster was restarted and a command hasn't been executed yet, or because a change was made that removed the command from a service entirely.

Regardless, we now have a way to visually indicate when this happens.